+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.1.1
+- refactoring of test case
+  - Introduced capture_stdout
+  - Introduced FakeFS
+- Change logdev of Console for Logger from STDOUT to $ stdout
+
 # 2.1.0
 - Integrate TeeLogger::Base class to TeeLogger::TeeLogger class